CrewAI/requirements.txt
streamlit
openai
python-dotenv
crewai
crewai-tools
